baselines/EMNLP2019/requirements.txt
neuralcoref==4.0
spacy==2.1.3
tensorflow==1.13.1
OpenNMT-tf==1.22.0
networkx==2.3
requests==2.21.0
sqlitedict==1.6.0
SPARQLWrapper==1.8.4
bert-tensorflow==1.0.1
tensorflow-hub==0.5.0
diffbot-nlapi==0.0.3